Code clenup. Fix review comments. Add checking for support Async Storage mode.

+++ infrastructures/kubernetes/src/main/java/org/eclipse/che/workspace/infrastructure/kubernetes/provision/AsyncStoragePodWatcher.java
@@ -10,17 +10,6 @@
  *   Red Hat, Inc. - initial API and implementation
  */
 
-/**
- * Periodically checks ability to stop Asynchronous Storage Pod. It will periodically revise
- * UserPreferences of all registered user and check specialized preferences. Preferences should be
- * recorded if last workspace stopped and cleanup on start any workspace. Required preferences to
- * initiate Asynchronous Storage Pod: {@link
- * org.eclipse.che.api.workspace.shared.Constants.LAST_ACTIVE_INFRASTRUCTURE_NAMESPACE} : should
- * contain last used infrastructure namespace {@link
- * org.eclipse.che.api.workspace.shared.Constants.LAST_ACTIVE_WORKSPACE_ID} : should contain last
- * used workspace id {@link org.eclipse.che.api.workspace.shared.Constants.LAST_ACTIVITY_TIME} :
- * seconds then workspace stopped in the Java epoch format
- */
 package org.eclipse.che.workspace.infrastructure.kubernetes.provision;
 
 import static com.google.common.base.Strings.isNullOrEmpty;
@@ -28,6 +17,7 @@ import static java.lang.Long.parseLong;
 import static org.eclipse.che.api.core.Pages.iterateLazily;
 import static org.eclipse.che.api.workspace.shared.Constants.LAST_ACTIVE_INFRASTRUCTURE_NAMESPACE;
 import static org.eclipse.che.api.workspace.shared.Constants.LAST_ACTIVITY_TIME;
+import static org.eclipse.che.workspace.infrastructure.kubernetes.namespace.pvc.CommonPVCStrategy.COMMON_STRATEGY;
 import static org.eclipse.che.workspace.infrastructure.kubernetes.provision.AsyncStorageProvisioner.ASYNC_STORAGE;
 
 import io.fabric8.kubernetes.api.model.DoneablePod;
@@ -47,11 +37,22 @@ import org.eclipse.che.api.user.server.UserManager;
 import org.eclipse.che.api.workspace.server.WorkspaceRuntimes;
 import org.eclipse.che.api.workspace.server.spi.InfrastructureException;
 import org.eclipse.che.api.workspace.server.spi.InternalRuntime;
+import org.eclipse.che.commons.annotation.Nullable;
 import org.eclipse.che.commons.schedule.ScheduleDelay;
 import org.eclipse.che.workspace.infrastructure.kubernetes.KubernetesClientFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * Periodically checks ability to stop Asynchronous Storage Pod. It will periodically revise
+ * UserPreferences of all registered user and check specialized preferences. Preferences should be
+ * recorded if last workspace stopped and cleanup on start any workspace. Required preferences to
+ * initiate stop procedure for Asynchronous Storage Pod : {@link
+ * org.eclipse.che.api.workspace.shared.Constants.LAST_ACTIVE_INFRASTRUCTURE_NAMESPACE} : should
+ * contain last used infrastructure namespace {@link
+ * {@link org.eclipse.che.api.workspace.shared.Constants.LAST_ACTIVITY_TIME} :
+ * seconds then workspace stopped in the Java epoch format
+ */
 @Singleton
 public class AsyncStoragePodWatcher {
 
@@ -63,6 +64,7 @@ public class AsyncStoragePodWatcher {
   private final WorkspaceRuntimes runtimes;
   private final long shutdownTimeoutSec;
   private final Clock clock = Clock.systemDefaultZone();
+  private final boolean isAsyncStoragePodCanBeRun;
 
   @Inject
   public AsyncStoragePodWatcher(
@@ -70,12 +72,44 @@ public class AsyncStoragePodWatcher {
       UserManager userManager,
       PreferenceManager preferenceManager,
       WorkspaceRuntimes runtimes,
-      @Named("che.infra.kubernetes.async.storage.shutdown_timeout_min") long shutdownTimeoutMin) {
+      @Named("che.infra.kubernetes.async.storage.shutdown_timeout_min") long shutdownTimeoutMin,
+      @Named("che.infra.kubernetes.pvc.strategy") String pvcStrategy,
+      @Named("che.infra.kubernetes.namespace.allow_user_defined")
+          boolean allowUserDefinedNamespaces,
+      @Nullable @Named("che.infra.kubernetes.namespace.default") String defaultNamespaceName,
+      @Named("che.limits.user.workspaces.run.count") int runtimesPerUser) {
     this.kubernetesClientFactory = kubernetesClientFactory;
     this.userManager = userManager;
     this.preferenceManager = preferenceManager;
     this.runtimes = runtimes;
-    this.shutdownTimeoutSec = shutdownTimeoutMin * 60;
+    this.shutdownTimeoutSec = TimeUnit.MINUTES.toSeconds(shutdownTimeoutMin);
+
+    isAsyncStoragePodCanBeRun =
+        isAsyncStoragePodCanBeRun(
+            pvcStrategy, allowUserDefinedNamespaces, defaultNamespaceName, runtimesPerUser);
+  }
+
+  /**
+   * Checking current system configuration on ability to run Async Storage Pod. Will be checked next
+   * value of properties:
+   *
+   * <ul>
+   *   <li>che.infra.kubernetes.namespace.default=<username>-che
+   *   <li>che.infra.kubernetes.namespace.allow_user_defined=false
+   *   <li>che.infra.kubernetes.pvc.strategy=common
+   *   <li>che.limits.user.workspaces.run.count=1
+   * </ul>
+   */
+  private boolean isAsyncStoragePodCanBeRun(
+      String pvcStrategy,
+      boolean allowUserDefinedNamespaces,
+      String defaultNamespaceName,
+      int runtimesPerUser) {
+    return !allowUserDefinedNamespaces
+        && COMMON_STRATEGY.equals(pvcStrategy)
+        && runtimesPerUser == 1
+        && !isNullOrEmpty(defaultNamespaceName)
+        && defaultNamespaceName.contains("<username>");
   }
 
   @ScheduleDelay(
@@ -83,33 +117,35 @@ public class AsyncStoragePodWatcher {
       initialDelay = 1,
       delayParameterName = "che.infra.kubernetes.async.storage.shutdown_check_period_min")
   public void check() {
-    for (User user :
-        iterateLazily((maxItems, skipCount) -> userManager.getAll(maxItems, skipCount))) {
-      try {
-        String owner = user.getId();
-        Map<String, String> preferences = preferenceManager.find(owner);
-        String lastTimeAccess = preferences.get(LAST_ACTIVITY_TIME);
-        String namespace = preferences.get(LAST_ACTIVE_INFRASTRUCTURE_NAMESPACE);
-        if (isNullOrEmpty(namespace)
-            || isNullOrEmpty(lastTimeAccess)
-            || isAnyRuntimeInProgress(owner)) {
-          continue;
-        }
-        long lastTimeAccessSec = parseLong(lastTimeAccess);
-        long epochSec = clock.instant().getEpochSecond();
-        if (epochSec - lastTimeAccessSec >= shutdownTimeoutSec) {
-          PodResource<Pod, DoneablePod> podDoneablePodPodResource =
-              kubernetesClientFactory
-                  .create()
-                  .pods()
-                  .inNamespace(namespace)
-                  .withName(ASYNC_STORAGE);
-          if (podDoneablePodPodResource.get() != null) {
-            podDoneablePodPodResource.delete();
+    if (isAsyncStoragePodCanBeRun) { // if system not support async storage mode do nothing
+      for (User user :
+          iterateLazily((maxItems, skipCount) -> userManager.getAll(maxItems, skipCount))) {
+        try {
+          String owner = user.getId();
+          Map<String, String> preferences = preferenceManager.find(owner);
+          String lastTimeAccess = preferences.get(LAST_ACTIVITY_TIME);
+          String namespace = preferences.get(LAST_ACTIVE_INFRASTRUCTURE_NAMESPACE);
+          if (isNullOrEmpty(namespace)
+              || isNullOrEmpty(lastTimeAccess)
+              || isAnyRuntimeInProgress(owner)) {
+            continue;
           }
+          long lastTimeAccessSec = parseLong(lastTimeAccess);
+          long epochSec = clock.instant().getEpochSecond();
+          if (epochSec - lastTimeAccessSec >= shutdownTimeoutSec) {
+            PodResource<Pod, DoneablePod> podDoneablePodPodResource =
+                kubernetesClientFactory
+                    .create()
+                    .pods()
+                    .inNamespace(namespace)
+                    .withName(ASYNC_STORAGE);
+            if (podDoneablePodPodResource.get() != null) {
+              podDoneablePodPodResource.delete();
+            }
+          }
+        } catch (InfrastructureException | ServerException e) {
+          LOG.error(e.getMessage(), e);
         }
-      } catch (InfrastructureException | ServerException e) {
-        LOG.error(e.getMessage(), e);
       }
     }
   }
